refactor(traffic_light_multi_camera_fusion): extract core logic from node (#12610)
* test(autoware_traffic_light_multi_camera_fusion): use PascalCase for test names

Rename gtest TestSuite and TestName arguments in test_utils.cpp to
PascalCase as required by Google Test naming conventions.


* refactor(autoware_traffic_light_multi_camera_fusion): rename node files to multi_camera_fusion

Pure rename in preparation for extracting the core fusion logic into a
separate class. The Node class still lives in the renamed files; only
include paths, include guards, and CMakeLists/test references are
updated to keep the build green. The logical separation between Node
and core class happens in the follow-up commit.


* refactor(autoware_traffic_light_multi_camera_fusion): extract MultiCameraFusion logic

Move the Node-specific code (rclcpp subscribers/publisher, diagnostics,
parameter declarations) out of multi_camera_fusion.{cpp,hpp} into a new
MultiCameraFusionNode in traffic_light_multi_camera_fusion_node.{cpp,hpp}.
What remains in multi_camera_fusion.{cpp,hpp} is a pure-logic class with
a fuse() entry point that takes ROS2 messages and returns a Result,
without depending on rclcpp::Node or rclcpp::Logger. Warnings are
returned via Result fields so the Node logs them (Error-as-Value).


* fix(autoware_traffic_light_multi_camera_fusion): satisfy cppcheck functionStatic/functionConst

Mark MultiCameraFusion::convert_output_msg and update_best_record as
static, and update_log_odds / update_group_info_for_element as const,
so that cppcheck --enable=all --inconclusive passes. These functions
do not modify the object state; the qualifiers reflect existing
behavior.


* refactor(autoware_traffic_light_multi_camera_fusion): return unmapped traffic light IDs as data

Previously the core MultiCameraFusion class built warning log strings
("Found Traffic Light Id = X which is not defined in Map") and returned
them via MultiCameraFusionResult::warnings. Replace this with a raw
vector of traffic light IDs (unmapped_traffic_light_ids) and move the
string formatting to MultiCameraFusionNode, so that the core class only
deals with data and presentation concerns stay in the Node.


---------
